+++ packages/gatsby/src/bootstrap/index.js
@@ -22,6 +22,8 @@ const tracer = require(`opentracing`).globalTracer()
 const preferDefault = require(`./prefer-default`)
 const nodeTracking = require(`../db/node-tracking`)
 const withResolverContext = require(`../schema/context`)
+const stackTrace = require(`stack-trace`)
+import errorParser from "../query/error-parser"
 // Add `util.promisify` polyfill for old node versions
 require(`util.promisify/shim`)()
 
@@ -78,13 +80,17 @@ module.exports = async (args: BootstrapArgs) => {
 
   // theme gatsby configs can be functions or objects
   if (config && config.__experimentalThemes) {
-    const themes = await loadThemes(config)
+    // TODO: deprecation message for old __experimentalThemes
+    const themes = await loadThemes(config, { useLegacyThemes: true })
     config = themes.config
 
     store.dispatch({
       type: `SET_RESOLVED_THEMES`,
       payload: themes.themes,
     })
+  } else if (config) {
+    const plugins = await loadThemes(config, { useLegacyThemes: false })
+    config = plugins.config
   }
 
   if (config && config.polyfill) {
@@ -391,7 +397,37 @@ module.exports = async (args: BootstrapArgs) => {
       context,
       withResolverContext(context, schema),
       context
-    )
+    ).then(result => {
+      if (result.errors) {
+        report.panicOnBuild(
+          result.errors
+            .map(e => {
+              // Find the file where graphql was called.
+              const file = stackTrace
+                .parse(e)
+                .find(file => /createPages/.test(file.functionName))
+              if (file) {
+                const structuredError = errorParser({
+                  message: e.message,
+                  location: {
+                    start: { line: file.lineNumber, column: file.columnNumber },
+                  },
+                  filePath: file.fileName,
+                })
+                structuredError.context = {
+                  ...structuredError.context,
+                  fromGraphQLFunction: true,
+                }
+                return structuredError
+              }
+              return null
+            })
+            .filter(_.isObject)
+        )
+      }
+
+      return result
+    })
   }
 
   // Collect pages.