+++ themes/gatsby-theme-blog/src/components/layout.js
@@ -1,72 +1,21 @@
-/* eslint-disable */
-import React from 'react'
-import { Link } from 'gatsby'
+import React from "react"
+import { css, Styled } from "theme-ui"
+import Header from "./header"
 
-import { rhythm, scale } from '../utils/typography'
-
-class Template extends React.Component {
-  render() {
-    const { location, children } = this.props
-    const rootPath = `${__PATH_PREFIX__}/`
-    let header
-
-    if (location.pathname === rootPath) {
-      header = (
-        <h1
-          css={{
-            ...scale(1.5),
-            marginBottom: rhythm(1.5),
-            marginTop: 0,
-          }}
-        >
-          <Link
-            css={{
-              boxShadow: 'none',
-              textDecoration: 'none',
-              color: 'inherit',
-            }}
-            to={'/'}
-          >
-            Gatsby Theme Blog
-          </Link>
-        </h1>
-      )
-    } else {
-      header = (
-        <h3
-          css={{
-            fontFamily: 'Montserrat, sans-serif',
-            marginTop: 0,
-            marginBottom: rhythm(-1),
-          }}
-        >
-          <Link
-            css={{
-              boxShadow: 'none',
-              textDecoration: 'none',
-              color: 'inherit',
-            }}
-            to={'/'}
-          >
-            Gatsby Theme Blog
-          </Link>
-        </h3>
-      )
-    }
-    return (
+export default ({ children, ...props }) => (
+  <Styled.root>
+    <Header {...props} />
+    <div>
       <div
-        css={{
-          marginLeft: 'auto',
-          marginRight: 'auto',
-          maxWidth: rhythm(24),
-          padding: `${rhythm(1.5)} ${rhythm(3 / 4)}`,
-        }}
+        css={css({
+          maxWidth: `container`,
+          mx: `auto`,
+          px: 3,
+          py: 4,
+        })}
       >
-        {header}
         {children}
       </div>
-    )
-  }
-}
-
-export default Template
+    </div>
+  </Styled.root>
+)