+++ themes/gatsby-theme-notes/src/components/notes.js
@@ -5,8 +5,14 @@ import FileList from "./file-list"
 import Breadcrumbs from "./breadcrumbs"
 import Layout from "./layout"
 
-export default ({ directories, files, breadcrumbs = [] }) => (
-  <Layout>
+export default ({
+  directories,
+  files,
+  breadcrumbs = [],
+  siteTitle,
+  ...props
+}) => (
+  <Layout {...props} title={siteTitle}>
     {breadcrumbs.length ? <Breadcrumbs links={breadcrumbs} /> : null}
     <DirectoryList directories={directories} />
     <FileList files={files} />